# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><font color="red"><b>Wichtiger allgemeiner Hinweis:</b> Beachten Sie unbedingt die Benennung der Form-Parameter, Query-Parameter sowie die der Ergebnistupel. Bei Abweichungen von in der API angegebenen Benennungen wird der entsprechende Endpunkt als <u><b>falsch</b></u> bewertet.</font></li>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
...
  /spiele:
    get:
      tags:
         - Anwender
      summary: Gibt Spiele zurück.
      description: 
        Alle **Spiele** werden zurückgegeben. Falls Query-Parameter mit Werten angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: angeschaut
          in: query
          description: 'Nur Spiele, die von mindestens einem Premium-Nutzer angeschaut (**angeschaut** == `true`) bzw. nicht angeschaut (**angeschaut** == `false`) wurden, werden zurückgegeben.'
          schema:
            type: boolean
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        spielid:
                          type: integer
                          description: '**RowID des Spiels**.'
                      required:
                        - spielid
                    - $ref: '#/components/schemas/spiel'
    post:
      tags:
        - Authentifizierter und als Mitarbeiter authorisierter Anwender
      summary: Ein Spiel wird hinzugefügt.
      description: 
        Ein **Spiel** wird hinzugefügt.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/spiel'
                - type: object
                  properties:
                    manschaft1id:
                      type: integer
                      description: '**RowID** der ersten Mannschaft.'
                    manschaft2id:
                      type: integer
                      description: '**RowID** der zweiten (gegnerischen) Mannschaft.'
                    ligaid:
                      type: integer
                      description: '**RowID** der zugeordneten Liga.'
                  required:
                    - manschaft1id
                    - manschaft2id
                    - ligaid
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403' 
...
  /ligen/{ligaid}/spiele:
    get:
      tags:
        - Anwender
      summary: Gibt Spiele einer Liga zurück.
      description: 
        Alle **Spiele** einer **Liga** werden zurückgegeben.
      parameters:
        - name: ligaid
          in: path
          description: '**RowID** der Liga.'
          schema:
           type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        spielid:
                          type: integer
                          description: '**RowID** des Spiels.'
                      required:
                        - verletzungid
                    - $ref: '#/components/schemas/spiel'
        '404':
          $ref: '#/components/responses/404'
  /spieler:
    get:
      tags:
         - Anwender
      summary: Gibt Spieler zurück.
      description: 
        Alle **Spieler** werden zurückgegeben. Falls Query-Parameter mit Werten angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: nachname
          in: query
          description: 'Nur Spieler, deren **Nachname** `nachname` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        spielerid:
                          type: integer
                          description: '**RowID des Spielers**.'
                      required:
                        - spielerid
                    - $ref: '#/components/schemas/spieler'
...
  /spieler/{spielerid}/verletzungen:
    get:
      tags:
        - Anwender
      summary: Gibt die Verletzungen eines Spielers zurück.
      description: 
        Alle **Verletzungen** von einem **Spieler** werden zurückgegeben.
      parameters:
        - name: spielerid
          in: path
          description: '**RowID** des Spielers.'
          schema:
           type: integer
          required: true
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        verletzungid:
                          type: integer
                          description: '**RowID** der Verletzung.'
                      required:
                        - verletzungid
                    - $ref: '#/components/schemas/verletzung'
        '404':
          $ref: '#/components/responses/404'
    post:
      tags:
        - Authentifizierter und als Mitarbeiter authorisierter Anwender
      summary: Fügt eine Verletzung zu einem Spieler hinzu.
      security:
        - basicAuth: []
      description: Eine **Verletzung** wird zu einem Spieler hinzugefügt.
      parameters:
      - name: spielerid
        in: path
        description: '**RowID** der Komposition.'
        schema:
          type: integer
        required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/verletzung'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
...
    spiel:
      type: object
      properties:
        link:
          type: string
          description: '**Link** des Spiels.'
        datum:
          type: string
          format: date
          description: '**Datum** des Spiels (in der Form "YYYY-MM-DD").'
      required:
        - link
        - datum
    spieler:
      type: object
      properties:
        vorname:
          type: string
          description: '**Vorname** des Spielers.'
        nachname:
          type: string
          description: '**Nachname** des Spielers.'
        spitzname:
          type: string
          description: '**Spitzname** des Spielers.'
      required:
        - vorname
        - nachname
...
    verletzung:
      type: object
      properties:
        art:
          type: string
          description: '**Art** der Verletzung.'
        dauer:
          type: integer
          description: '**Dauer** der Verletzung.'
      required:
        - art
        - dauer
...